refactor(project-map): 收口大文件拆分
