+++ src/styles/project-map.evidence-files.css
@@ -1,0 +1,242 @@
+.project-map-evidence-files-panel {
+  display: grid;
+  gap: 10px;
+  max-height: min(36vh, 420px);
+  padding: 10px;
+  border: 1px solid var(--project-map-border);
+  border-radius: 16px;
+  background: color-mix(in srgb, var(--project-map-surface) 88%, transparent);
+  box-shadow: 0 14px 40px color-mix(in srgb, var(--project-map-shadow) 8%, transparent);
+  overflow: auto;
+}
+
+.project-map-evidence-files-panel.is-collapsed {
+  max-height: none;
+  overflow: hidden;
+}
+
+.project-map-evidence-files-header {
+  display: grid;
+  grid-template-columns: auto minmax(0, 1fr) auto;
+  gap: 9px;
+  align-items: start;
+}
+
+.project-map-evidence-files-header button {
+  height: 26px;
+  border: 1px solid var(--project-map-border);
+  border-radius: 999px;
+  background: var(--project-map-control);
+  color: var(--project-map-text);
+  font-size: 10px;
+}
+
+.project-map-evidence-files-header > svg {
+  width: 16px;
+  height: 16px;
+  color: var(--project-map-accent);
+}
+
+.project-map-evidence-files-header h4,
+.project-map-evidence-file-detail h5 {
+  margin: 0;
+  color: var(--project-map-text);
+  font-size: 12px;
+  line-height: 1.25;
+}
+
+.project-map-evidence-files-header p,
+.project-map-evidence-file-detail p,
+.project-map-evidence-files-empty,
+.project-map-evidence-non-file {
+  margin: 3px 0 0;
+  color: var(--project-map-text-muted);
+  font-size: 10px;
+  line-height: 1.35;
+}
+
+.project-map-evidence-files-controls {
+  display: grid;
+  grid-template-columns: minmax(160px, 1fr) minmax(112px, auto) auto;
+  gap: 7px;
+  align-items: center;
+}
+
+.project-map-evidence-files-controls select {
+  min-width: 0;
+  height: 28px;
+  border: 1px solid var(--project-map-border);
+  border-radius: 10px;
+  background: var(--project-map-control);
+  color: var(--project-map-text);
+  font-size: 10px;
+}
+
+.project-map-evidence-files-toggle {
+  display: inline-flex;
+  gap: 5px;
+  align-items: center;
+  color: var(--project-map-text-muted);
+  font-size: 10px;
+  white-space: nowrap;
+}
+
+.project-map-evidence-file-list {
+  display: grid;
+  gap: 6px;
+  max-height: 220px;
+  overflow: auto;
+}
+
+.project-map-evidence-file-row {
+  display: grid;
+  grid-template-columns: minmax(0, 1fr) auto;
+  gap: 3px 8px;
+  align-items: center;
+  width: 100%;
+  padding: 7px 8px;
+  border: 1px solid var(--project-map-border);
+  border-radius: 12px;
+  background: color-mix(in srgb, var(--project-map-control) 86%, transparent);
+  color: var(--project-map-text);
+  text-align: left;
+}
+
+.project-map-evidence-file-row:hover,
+.project-map-evidence-file-row:focus-visible,
+.project-map-evidence-file-row.is-selected {
+  border-color: color-mix(in srgb, var(--project-map-accent) 56%, var(--project-map-border));
+  background: color-mix(in srgb, var(--project-map-accent) 10%, var(--project-map-control));
+}
+
+.project-map-evidence-file-path {
+  overflow: hidden;
+  font-size: 11px;
+  font-weight: 650;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.project-map-evidence-file-meta {
+  color: var(--project-map-text-muted);
+  font-size: 10px;
+  white-space: nowrap;
+}
+
+.project-map-evidence-file-tags {
+  display: flex;
+  grid-column: 1 / -1;
+  gap: 4px;
+  flex-wrap: wrap;
+}
+
+.project-map-evidence-file-tags em {
+  padding: 2px 5px;
+  border-radius: 999px;
+  background: color-mix(in srgb, var(--project-map-accent) 10%, transparent);
+  color: var(--project-map-text-muted);
+  font-size: 9px;
+  font-style: normal;
+}
+
+.project-map-evidence-file-tags em.is-warning {
+  background: color-mix(in srgb, var(--project-map-warning) 14%, transparent);
+  color: var(--project-map-warning);
+}
+
+.project-map-evidence-file-tags em.is-changed {
+  background: color-mix(in srgb, var(--project-map-accent) 16%, transparent);
+  color: var(--project-map-accent);
+}
+
+.project-map-evidence-file-tags em.is-degraded {
+  background: color-mix(in srgb, var(--project-map-error) 10%, transparent);
+  color: var(--project-map-error);
+}
+
+.project-map-evidence-file-detail {
+  display: grid;
+  gap: 9px;
+  padding: 9px;
+  border: 1px solid color-mix(in srgb, var(--project-map-accent) 28%, var(--project-map-border));
+  border-radius: 14px;
+  background: color-mix(in srgb, var(--project-map-accent) 7%, transparent);
+}
+
+.project-map-evidence-file-detail header {
+  display: flex;
+  gap: 10px;
+  align-items: flex-start;
+  justify-content: space-between;
+}
+
+.project-map-evidence-file-actions,
+.project-map-evidence-related-nodes {
+  display: flex;
+  gap: 6px;
+  flex-wrap: wrap;
+}
+
+.project-map-evidence-file-actions button,
+.project-map-evidence-related-nodes button {
+  border: 1px solid var(--project-map-border);
+  border-radius: 999px;
+  background: var(--project-map-control);
+  color: var(--project-map-text);
+  font-size: 10px;
+}
+
+.project-map-evidence-file-actions button:disabled {
+  opacity: 0.5;
+}
+
+.project-map-evidence-related-nodes {
+  display: grid;
+}
+
+.project-map-evidence-related-nodes strong,
+.project-map-evidence-line-refs strong {
+  color: var(--project-map-text);
+  font-size: 10px;
+}
+
+.project-map-evidence-related-nodes button {
+  display: flex;
+  justify-content: space-between;
+  border-radius: 10px;
+  text-align: left;
+}
+
+.project-map-evidence-related-nodes button span {
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.project-map-evidence-related-nodes button em {
+  color: var(--project-map-text-muted);
+  font-style: normal;
+  white-space: nowrap;
+}
+
+.project-map-evidence-line-refs {
+  display: grid;
+  gap: 3px;
+  color: var(--project-map-text-muted);
+  font-size: 10px;
+}
+
+.project-map-node.is-evidence-file-node {
+  border-color: color-mix(in srgb, var(--project-map-accent) 72%, var(--project-map-border));
+  box-shadow: 0 0 0 3px color-mix(in srgb, var(--project-map-accent) 18%, transparent), 0 20px 40px color-mix(in srgb, var(--project-map-core) 16%, transparent);
+}
+
+@media (max-width: 900px) {
+  .project-map-evidence-files-controls {
+    grid-template-columns: 1fr;
+  }
+
+  .project-map-evidence-file-detail header {
+    display: grid;
+  }
+}
